+++ src/main/kotlin/com/pravo/pravo/global/error/ErrorResponse.kt
@@ -1,4 +1,4 @@
-package com.pravo.pravo.global.common.error
+package com.pravo.pravo.global.error
 
 import io.swagger.v3.oas.annotations.media.Schema
