+++ src/main/kotlin/com/pravo/pravo/global/error/exception/BaseException.java
@@ -1,6 +1,7 @@
-package com.pravo.pravo.global.common.error.exception;
+package com.pravo.pravo.global.error.exception;
 
-import com.pravo.pravo.global.common.error.ErrorCode;
+import com.pravo.pravo.global.error.ErrorCode;
+import com.pravo.pravo.global.error.ErrorCode;
 
 public class BaseException extends RuntimeException {
     private final ErrorCode errorCode;
